Data.SOP.BasicFunctors

Description

Basic functors.

Definitions of the type-level equivalents of const, id, and (.), and a definition of the lifted function space.

These datatypes are generally useful, but in this library, they're primarily used as parameters for the NP, NS, POP, and SOP types.

We define own variants of Const, Identity and Compose for various reasons.

Const and Compose become kind polymorphic only in base-4.9.0.0 (transformers-0.5.0.0).

Shorter names are convenient, and pattern synonyms aren't (yet) powerful enough, particularly exhaustiveness check doesn't work properly. See https://ghc.haskell.org/trac/ghc/ticket/8779.

Basic functors

newtype K (a :: Type) (b :: k)

The constant type functor.

Like Constant, but kind-polymorphic in its second argument and with a shorter name.

Constructors

K a

unK :: K a b -> a

Extract the contents of a K value.

newtype I (a :: Type)

The identity type functor.

Like Identity, but with a shorter name.

Constructors

I a

unI :: I a -> a

Extract the contents of an I value.

newtype ((f :: l -> Type) :.: (g :: k -> l)) (p :: k) infixr 7

Composition of functors.

Like Compose, but kind-polymorphic and with a shorter name.

Constructors

Comp (f (g p))

unComp :: (f :.: g) p -> f (g p)

Extract the contents of a Comp value.

Mapping functions

mapII :: (a -> b) -> I a -> I b

Lift the given function.

Since: 0.2.5.0

mapIK :: (a -> b) -> I a -> K b c

Lift the given function.

Since: 0.2.5.0

mapKI :: (a -> b) -> K a c -> I b

Lift the given function.

Since: 0.2.5.0

mapKK :: (a -> b) -> K a c -> K b d

Lift the given function.

Since: 0.2.5.0

mapIII :: (a -> b -> c) -> I a -> I b -> I c

Lift the given function.

Since: 0.2.5.0

mapIIK :: (a -> b -> c) -> I a -> I b -> K c d

Lift the given function.

Since: 0.2.5.0

mapIKI :: (a -> b -> c) -> I a -> K b d -> I c

Lift the given function.

Since: 0.2.5.0

mapIKK :: (a -> b -> c) -> I a -> K b d -> K c e

Lift the given function.

Since: 0.2.5.0

mapKII :: (a -> b -> c) -> K a d -> I b -> I c

Lift the given function.

Since: 0.2.5.0

mapKIK :: (a -> b -> c) -> K a d -> I b -> K c e

Lift the given function.

Since: 0.2.5.0

mapKKI :: (a -> b -> c) -> K a d -> K b e -> I c

Lift the given function.

Since: 0.2.5.0

mapKKK :: (a -> b -> c) -> K a d -> K b e -> K c f

Lift the given function.

Since: 0.2.5.0